requirements.txt
mxnet
lxml
requests
pillow
cv2